Refactor ShowRequest, extracting Req component outside
